Happiness Dataset

A data frame containing a happiness survey and demographic data. This data are fictitious.

happiness

Format

A data frame with 460 rows and 11 variables:

ID

character. A unique identifier.

Date

date. Date of the interview.

Sex

factor. Sex coded as Male or Female.

Race

factor. Race coded as an 8-level factor.

Age

integer. Age in years.

Education

factor. Education coded as a 13-level factor.

Income

double. Annual income in US dollars.

IQ

double. Adult intelligence quotient. This variable has a large amount of missing data.

Zip

character. USPS Zip code.

Children

integer. Number of children.

Happy

factor. Agreement with the statement "I am happy most of the time", coded as Strongly Disagree , Disagree, Neutral, Agree, or Strongly Agree.

Source

The data were randomly generated using functions from the wakefield package.